removed old test file, moved into main, and cleaned up
