+++ btf/include/LogPrinter.h
@@ -1,15 +1,18 @@
+#ifndef __LOG_PRINTER_H__
+#define __LOG_PRINTER_H__
+
 #include <stdio.h>
+#include <sys/ioctl.h>
+#include <unistd.h>
+
+#include <string>
 
 #include "gtest/gtest.h"
 
-using ::testing::EmptyTestEventListener;
-using ::testing::InitGoogleTest;
-using ::testing::Test;
-using ::testing::TestCase;
-using ::testing::TestEventListeners;
-using ::testing::TestInfo;
-using ::testing::TestPartResult;
-using ::testing::UnitTest;
+using testing::EmptyTestEventListener;
+using testing::TestCase;
+using testing::UnitTest;
+
 class LogPrinter : public EmptyTestEventListener {
   struct Log {
     std::string TestName;
@@ -23,3 +26,5 @@ class LogPrinter : public EmptyTestEventListener {
   virtual void OnTestStart(const testing::TestInfo &test_info);
   virtual void OnTestEnd(const testing::TestInfo &test_info);
 };
+
+#endif
